+++ cpskin/core/upgradehandlers.py
@@ -6,6 +6,7 @@ from cpskin.core.faceted.interfaces import ICPSkinPossibleFacetedNavigable
 from cpskin.core.setuphandlers import addAutoPlaySliderToRegistry
 from cpskin.core.setuphandlers import addCityNameToRegistry
 from cpskin.core.setuphandlers import addLoadPageMenuToRegistry
+from cpskin.core.setuphandlers import addMediaViewletOptionsToRegistry
 from cpskin.core.setuphandlers import addPortletsInRightActionsToRegistry
 from cpskin.core.setuphandlers import addShowSloganToRegistry
 from cpskin.core.setuphandlers import addSliderTimerToRegistry
@@ -27,6 +28,10 @@ import logging
 
 
 logger = logging.getLogger('cpskin.core')
+
+
+def upgrade_registry_for_media_viewlet(context):
+    addMediaViewletOptionsToRegistry(upgrade=True)
 
 
 def add_right_actions(context):